refactor: api 파일 이름 형식 통일
